Rename updater.h to better describe it's purpose
